Refactor image-selection
Change-type: patch
Changelog-entry: Use React instead of Angular for image selection
