+++ lib/gui/app/pages/main/controllers/main.js
@@ -31,7 +31,8 @@ const prettyBytes = require('pretty-bytes')
 module.exports = function (
   TooltipModalService,
   OSOpenExternalService,
-  $filter
+  $filter,
+  $scope
 ) {
   // Expose several modules to the template for convenience
   this.selection = selectionState
@@ -42,6 +43,13 @@ module.exports = function (
   this.constraints = driveConstraints
   this.progressMessage = messages.progress
   this.isWebviewShowing = Boolean(store.getState().toJS().isWebviewShowing)
+
+  // Trigger an update if the store changes
+  store.observe(() => {
+    if (!$scope.$$phase) {
+      $scope.$apply()
+    }
+  })
 
   /**
    * @summary Determine if the drive step should be disabled
