+++ lib/gui/app/pages/main/main.js
@@ -54,7 +54,6 @@ const MainPage = angular.module(MODULE_NAME, [
 ])
 
 MainPage.controller('MainController', require('./controllers/main'))
-MainPage.controller('ImageSelectionController', require('./controllers/image-selection'))
 MainPage.controller('DriveSelectionController', require('./controllers/drive-selection'))
 MainPage.controller('FlashController', require('./controllers/flash'))
